Update form/BusyButton to make it usable in AMD.
