+++ form/BusyComboButton.js
@@ -1,0 +1,8 @@
+define([
+	"./_BusyButtonMixin",
+	"dijit/form/ComboButton",
+	"dojo/_base/declare"
+], function(_BusyButtonMixin, ComboButton, declare){
+return declare("dojox.form.BusyComboButton", [ComboButton, _BusyButtonMixin], {});
+});
+
